modifying tutorial to fix errors
