+++ devon4j-http-rest-client/files/VisitormanagementRestService.java
@@ -1,4 +1,4 @@
-package com.example.application.httprestserver.visitormanagement.service.api.rest;
+package com.example.application.visitormanagement.service.api.rest;
 
 import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
@@ -6,14 +6,15 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
-
-@Path("/visitormanagement")
+@Path("/visitormanagement/v1")
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
 public interface VisitormanagementRestService {
 
   @GET
-  @Path("/clientrequest")
+  @Path("/clientrequest/")
+  @Consumes(MediaType.APPLICATION_JSON)
+  @Produces(MediaType.APPLICATION_JSON)
   public String returnResponseToClient();
 
 }
