+++ devon4j-http-rest-client/files/server_application.txt
@@ -3,7 +3,7 @@
 # If you are running in a servlet container you may add this to lib/config/application.properties in case you do not
 # want to touch the WAR file.
 
-server.port=8080
+server.port=8081
 server.servlet.context-path=/httprestserver
 spring.application.name=httprestserver
